Move sim plugin
Move the simulator plugin files to $(pythondir)/sim_plugin
https://github.com/libstorage/libstoragemgmt/issues/440
