+++ plugin/sim_plugin/Makefile.am
@@ -1,8 +1,7 @@
 if WITH_SIM
 
-plugindir = $(pythondir)/lsm/plugin
+simdir = $(pythondir)/sim_plugin
 
-simdir = $(plugindir)/sim
 sim_PYTHON = \
 	__init__.py \
 	simulator.py \
@@ -11,7 +10,5 @@ sim_PYTHON = \
 dist_bin_SCRIPTS= sim_lsmplugin
 
 EXTRA_DIST = sim_lsmplugin.in
-
-
 
 endif
