+++ plugin/sim_plugin/__init__.py
@@ -1,0 +1,1 @@
+from sim_plugin.simulator import SimPlugin
